// example/prisma/generated/prisma-select/schema.ts
// Generated by prisma-select - DO NOT EDIT
import type { Schema } from 'prisma-select';

export const schema = {
  User: {
    fields: {
      id: { kind: 'scalar', type: 'Int' },
      email: { kind: 'scalar', type: 'String' },
      name: { kind: 'scalar', type: 'String' },
      posts: { kind: 'object', type: 'Post' }
    }
  },
  Post: {
    fields: {
      id: { kind: 'scalar', type: 'Int' },
      title: { kind: 'scalar', type: 'String' },
      body: { kind: 'scalar', type: 'String' },
      published: { kind: 'scalar', type: 'Boolean' },
      author: { kind: 'object', type: 'User' },
      authorId: { kind: 'scalar', type: 'Int' },
      comments: { kind: 'object', type: 'Comment' }
    }
  },
  Comment: {
    fields: {
      id: { kind: 'scalar', type: 'Int' },
      text: { kind: 'scalar', type: 'String' },
      post: { kind: 'object', type: 'Post' },
      postId: { kind: 'scalar', type: 'Int' }
    }
  }
} as const satisfies Schema;

export type ModelName = 'User' | 'Post' | 'Comment';

export interface ModelsObject {
  User: { id: true; email: true; name: true; posts: true; _count: true };
  Post: { id: true; title: true; body: true; published: true; author: true; authorId: true; comments: true; _count: true };
  Comment: { id: true; text: true; post: true; postId: true; _count: true };
}